+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.0.1 [2022.05.07]
 
 - [QueryBuilder] Added `selectDistinct()` method.
+- [QueryBuilder] Added `regexp()`, `andRegexp()`, `orRegexp()` methods.